alloc code done

+++ MP4/MP4_Sources/MP4_Sources/vm_pool.C
@@ -33,7 +33,7 @@ VMPool::VMPool(unsigned long  _base_address,
 
     /*_______ register new VMPool_______*/
     this->_page_table->register_pool(this);
-    
+
     /*_______initialize instance variables_______*/
     this->_base_address = _base_address;
     this->_size         = _size;
@@ -129,8 +129,8 @@ void VMPool::release(unsigned long _start_address) {
 
 bool VMPool::is_legitimate(unsigned long _address) {
 
-                Console::puts("         -> is_legitimate: base_address     =");Console::puti((unsigned int) this->_base_address);Console::puts("\n");
-                Console::puts("         -> is_legitimate: checking address =");Console::puti((unsigned int) _address);Console::puts("\n");
+                Console::puts("         -> is_legitimate: base_address     =");Console::puti(this->_base_address);Console::puts("\n");
+                Console::puts("         -> is_legitimate: checking address =");Console::puti(_address);Console::puts("\n");
 
     // if address belongs to free/alloc lists' region
     if((this->_base_address <= _address) && ((this->_base_address + 2*PAGE_SIZE) > _address)){
